refactor: 클래스 이름 변경 및 구조 변경
