+++ src/main/java/com/goormcoder/ieum/domain/Destination.java
@@ -1,5 +1,6 @@
 package com.goormcoder.ieum.domain;
 
+import com.goormcoder.ieum.domain.enumeration.DestinationName;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
@@ -16,20 +17,20 @@ public class Destination {
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, unique = true)
-    private DestinationType destinationType;
+    private DestinationName destinationName;
 
     @Column(nullable = false)
     private String description;
 
     @Builder
-    private Destination(DestinationType destinationType, String description) {
-        this.destinationType = destinationType;
+    private Destination(DestinationName destinationName, String description) {
+        this.destinationName = destinationName;
         this.description = description;
     }
 
-    public static Destination of(DestinationType destinationType, String description) {
+    public static Destination of(DestinationName destinationName, String description) {
         return Destination.builder()
-                .destinationType(destinationType)
+                .destinationName(destinationName)
                 .description(description)
                 .build();
     }
