+++ src/main/java/com/goormcoder/ieum/dto/request/PlanCreateDto.java
@@ -1,7 +1,6 @@
 package com.goormcoder.ieum.dto.request;
 
-import com.goormcoder.ieum.domain.PlanVehicle;
-import jakarta.validation.constraints.FutureOrPresent;
+import com.goormcoder.ieum.domain.enumeration.PlanVehicle;
 import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDateTime;
